Remove several externals that are more or less Linux specific from Windows build. Deferred fixing DRMAA for now due to the sticky build setup. Deferred fixing boost until the way it's getting shoved in can be cleaned up and a list of needed boost libraries is figured out. Made a couple of code fixes to support building 64bit version of HTCondor on Windows. Ticket #4022 Ticket #4023

+++ src/condor_startd.V6/Resource.cpp
@@ -1886,7 +1886,7 @@ Resource::publish( ClassAd* cap, amask_t mask )
 		case PARTITIONABLE_SLOT:
 			cap->AssignExpr(ATTR_SLOT_PARTITIONABLE, "TRUE");
             cap->Assign(ATTR_SLOT_TYPE, "Partitionable");
-            if (r_has_cp) cap->Assign(ATTR_NUM_CLAIMS, r_claims.size());
+            if (r_has_cp) cap->Assign(ATTR_NUM_CLAIMS, (long long)r_claims.size());
 			break;
 		case DYNAMIC_SLOT:
 			cap->AssignExpr(ATTR_SLOT_DYNAMIC, "TRUE");
@@ -2151,7 +2151,7 @@ Resource::publish_private( ClassAd *ad )
             claims += (*j)->id();
         }
         ad->Assign(ATTR_CLAIM_ID_LIST, claims);
-        ad->Assign(ATTR_NUM_CLAIMS, r_claims.size());
+        ad->Assign(ATTR_NUM_CLAIMS, (long long)r_claims.size());
     }
 }
